+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- Rename `on_day` method to `day_of_month`
 - Improve error messages by using the message containerd in the API response when
 available.